btf: Export and marshal DeclTags
This commit makes the DeclTag BTF type exported. The DeclTag has been
unexported because its encoding is a bit tricky.

DeclTags on the wire refer to a type by its ID, and an component index.
Unlike other BTF types, the type ID isn't always the actual "target",
instead if the component index is not `-1` it targets a child object of
the type. For unions and structs, the `Target` can be a member, for
functions it can be a FuncParam. Upon marshalling the correct component
index is calculated. This allows for modifying BTF types or creating
them without having to think about the component index.

In addition, the BTF marshalling logic that creates a minimal BTF blob
for the program has been updated to include DeclTags associated with
FuncInfos. Before we were not adding the DeclTags to the generated BTF
even if the ELF included them.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>